Accept decimal fraction number for -H.

+++ schaf.h
@@ -60,7 +60,7 @@ Value eval_string(const char *s);
 
 const char *error_message(void);
 
-void sch_set_gc_init_size(size_t mib);
+void sch_set_gc_init_size(double mib);
 void sch_set_gc_stress(bool b);
 
 #endif
